Working on parsing docx files figured out a way to get tables out of docx as well.
